+++ src/basic/components/CartPage.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useState } from 'react';
 
-import type { CartItem, Coupon, NotificationVariant } from '../../types';
+import type { Coupon, NotificationVariant } from '../../types';
 import type { ProductWithUI } from '../constants';
 import { Icon } from './icons';
 import { calculateCartTotal, getRemainingStock } from '../models/entity';
@@ -9,8 +9,8 @@ import { CartItemList } from './ui/CartItemList';
 import { CouponSelector } from './ui/CouponSelector';
 import { PaymentSummary } from './ui/PaymentSummary';
 import { ProductList } from './ui/ProductList';
+import { useCartStore } from '../hooks/useCartStore';
 import { useDebounce } from '../utils/hooks/useDebounce';
-import { useLocalStorage } from '../utils/hooks/useLocalStorage';
 
 interface CartPageProps {
   setIsAdmin: (isAdmin: boolean) => void;
@@ -35,10 +35,10 @@ export function CartPage({
 
   onAddNotification,
 }: CartPageProps) {
+  const { cart, addToCart, updateToCart, removeFromCart, resetCart } = useCartStore();
+
   const [searchTerm, setSearchTerm] = useState('');
   const debouncedSearchTerm = useDebounce(searchTerm, 500);
-
-  const [cart, setCart] = useLocalStorage<CartItem[]>('cart', []);
 
   const applyCoupon = useCallback(
     (coupon: Coupon) => {
@@ -58,11 +58,11 @@ export function CartPage({
   const completeOrder = useCallback(() => {
     const orderNumber = `ORD-${Date.now()}`;
     onAddNotification(`주문이 완료되었습니다. 주문번호: ${orderNumber}`, 'success');
-    setCart([]);
+    resetCart();
     setSelectedCoupon(null);
-  }, [onAddNotification]);
+  }, [onAddNotification, resetCart]);
 
-  const addToCart = useCallback(
+  const handleAddToCart = useCallback(
     (product: ProductWithUI) => {
       const remainingStock = getRemainingStock(product, cart);
       if (remainingStock <= 0) {
@@ -70,33 +70,25 @@ export function CartPage({
         return;
       }
 
-      setCart((prevCart) => {
-        const existingItem = prevCart.find((item) => item.product.id === product.id);
+      const existingItem = cart.find((item) => item.product.id === product.id);
 
-        if (existingItem) {
-          const newQuantity = existingItem.quantity + 1;
+      if (existingItem) {
+        const newQuantity = existingItem.quantity + 1;
 
-          if (newQuantity > product.stock) {
-            onAddNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
-            return prevCart;
-          }
-
-          return prevCart.map((item) =>
-            item.product.id === product.id ? { ...item, quantity: newQuantity } : item
-          );
+        if (newQuantity > product.stock) {
+          onAddNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
+          return;
         }
 
-        return [...prevCart, { product, quantity: 1 }];
-      });
+        updateToCart(product, newQuantity);
+        return;
+      }
 
+      addToCart(product);
       onAddNotification('장바구니에 담았습니다', 'success');
     },
-    [cart, onAddNotification]
+    [cart, onAddNotification, addToCart, updateToCart]
   );
-
-  const removeFromCart = useCallback((productId: string) => {
-    setCart((prevCart) => prevCart.filter((item) => item.product.id !== productId));
-  }, []);
 
   const updateQuantity = useCallback(
     (productId: string, newQuantity: number) => {
@@ -114,13 +106,9 @@ export function CartPage({
         return;
       }
 
-      setCart((prevCart) =>
-        prevCart.map((item) =>
-          item.product.id === productId ? { ...item, quantity: newQuantity } : item
-        )
-      );
+      updateToCart(product, newQuantity);
     },
-    [products, removeFromCart, onAddNotification, getRemainingStock]
+    [products, removeFromCart, onAddNotification, updateToCart]
   );
 
   const totals = calculateCartTotal(cart, selectedCoupon);
@@ -147,7 +135,7 @@ export function CartPage({
                 products={products}
                 debouncedSearchTerm={debouncedSearchTerm}
                 cart={cart}
-                addToCart={addToCart}
+                addToCart={handleAddToCart}
               />
             </section>
           </div>
